% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.r
\name{dfm_dictionary}
\alias{dfm_dictionary}
\title{Combining a Dataframe with Dictionary Counts}
\usage{
dfm_dictionary(data, string_vector)
}
\arguments{
\item{data}{the data input}

\item{string_vector}{the dictionary object}
}
\value{
the combined dataframe with the dictionary counts as a new column to it
}
\description{
Combining a Dataframe with Dictionary Counts
}
